+++ tests/files/sub/test.php
@@ -1,0 +1,11 @@
+<?php
+
+return [
+    'this' => [
+        'is' => [
+            'a' => [
+                'test' => 'something'
+            ]
+        ]
+    ]
+];